Add --licenses flag to list all licenses/vars
This makes it easier to write tooling around Lice. For example, I would
like to write an emacs package at some point, but it'll need to know
the valid licenses and their arguments so that it can prompt for them
if necessary.

It's also just nice to be able to see which licenses are included.