[Test] Configure Travis CI to test with Python 3.4 and 3.5
This change increases the versions of Python with which we test lice
for any errors, just in case anything breaks if we use the more recent
releases of the language.  The patch also adds my address to the
Travis CI configuration so that I will receive any emails related to
failed builds.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>

Resolves: #40
Special-Thanks: Corey Farwell