[Fix] Remove the boilerplate from the GPL templates
The templates for both the GPLv2 and GPLv3 contain all of the text
describing how to apply each license to a program.  It is not
necessary to include this information in the actual licenses
themselves.  So this patch removes that boilerplate to both cut down
on file-size and to include only what is actually necessary for each
of the two licenses.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>

Resolves: #44