[Feature] Add support for the Rust language
We use the C-style for file-header block comments.  Often in Rust one
will see block comments at the top of files that begin with `//!`, but
such comments appear in auto-generated documentation.  So we use the
C-style under the assumption developers will not want the license
information to appear in the documentation for every file.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>