[Fix] Correct the template for the WTFPL
Our template for WTFPL inserts variables into the template,
specifcially for the year and organization.  However, the text for the
WTFPL should be included verbatim without any variables.  For that
license the year and organization refers directly to the author of the
WTFPL itself, and so we should not change that.  This patch removes
the variables from our WTFPL template so that lice will generate the
license with the proper information.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>

GitHub-Issue: #46
See-Also: http://www.wtfpl.net/faq/