[Feature] Add support for `--header wtfpl`
This patch adds a header version of the WTFPL so that users can use
WTFPL in conjunction with the `--header` command-line flag.  The text
for the header links to the WTFPL site instead of mentioning the
`COPYING` file (see the link below) because there is no guarantee that
users will put their copy of the WTFPL into a file named `COPYING`.
Instead of potentially naming a non-existent file we use the
alternative version of the header text that references the official
site for the license.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>

See-Also: http://www.wtfpl.net/faq/
Closes: #46